+++ src/modules/nodes/NodesApi.ts
@@ -1,14 +1,19 @@
-import { NodeType } from './NodesTypes';
+import { NodeInfoType, NodeType } from './NodesTypes';
 import { AxiosPromise } from 'axios';
 
 import API from '@/libs/request';
 
 const URLS = {
     LIST: '/node/',
+    INFO: '/info/',
 };
 
 export const listNodes = (): AxiosPromise<NodeType[]> => API.get(URLS.LIST);
 
+export const retrieveInfo = (): AxiosPromise<NodeInfoType> =>
+    API.get(URLS.INFO);
+
 export default {
     listNodes,
+    retrieveInfo,
 };